Add missing tests for chain_rule_derivative and clean up test_Calculus.py
