+++ Tests/test_Calculus.py
@@ -64,7 +64,6 @@ if math_dir not in sys.path:
 from Math.Calculus.Differentiation.product_rule import compute_polynomial_derivative_str, product_rule_derivative, format_polynomial as format_polynomial_product_rule
 from Math.Calculus.Integration.NumIntegration import integrate_polynomial, format_polynomial_integration
 from Math.Calculus.Differentiation.quotient_rule import format_polynomial, quotient_rule_derivative, compute_polynomial_derivative_str as compute_polynomial_derivative_str_quotient
-from Math.Calculus.Differentiation.chain_rule import chain_rule_derivative, format_polynomial as format_polynomial_chain_rule, compute_polynomial_derivative
 from Math.Calculus.Integration.TrigIntegration import integrate_cos, integrate_sin
 from Math.Calculus.Differentiation.simple_diffrentiation_function import differentiate_polynomial
 
@@ -378,145 +377,25 @@ def test_product_rule_derivative_both_constants():
     result = product_rule_derivative([5], [0], [7], [0])
     assert result == "(0) * (7x^0) + (5x^0) * (0)"
 
-def test_chain_rule_derivative_basic():
-    # g(x) = x^2, n = 3 => (x^2)^3 => derivative is 3(x^2)^2 * (2x)
-    result = chain_rule_derivative([1], [2], 3)
-    # Check parts of the string independently to avoid brittle tests
-    assert "3(" in result
-    assert "1x^2" in result
-    assert ")^2" in result
-    assert "* (2x^1)" in result
-
-def test_chain_rule_derivative_multiple_terms():
-    # g(x) = x^2 + 2x, n = 3
-    result = chain_rule_derivative([1, 2], [2, 1], 3)
-    assert "3(" in result
-    assert "1x^2 + 2x^1" in result
-    assert ")^2" in result
-    assert "* (2x^1 + 2x^0)" in result
-
-def test_chain_rule_derivative_constant_inner():
-    # g(x) = 5, n = 2 => (5)^2 => derivative is 2(5)^1 * 0
-    result = chain_rule_derivative([5], [0], 2)
-    assert "2(5x^0)^1" in result
-    assert "* ()" in result
-
-def test_chain_rule_derivative_zero_exponent():
-    # g(x) = x^2 + 2, n = 0 => (x^2 + 2)^0 => derivative is 0(...)
-    result = chain_rule_derivative([1, 2], [2, 0], 0)
-    assert result.startswith("0(")
-
-def test_chain_rule_derivative_negative_exponent():
-    # Negative exponent should raise ValueError
-    with pytest.raises(ValueError, match="Exponent must be non-negative."):
-        chain_rule_derivative([1, 2], [2, 0], -1)
 
 
-def test_chain_rule_derivative_float_coefficients():
-    # g(x) = 1.5x^2.5, n = 2 => 2(1.5x^2.5)^1 * (3.75x^1.5)
-    result = chain_rule_derivative([1.5], [2.5], 2)
-    assert "2(" in result
-    assert "1.5x^2" in result
-    assert ")^1" in result
-    assert "* (3.75x^1)" in result
 
-def test_chain_rule_derivative_empty():
-    # g(x) = 0 (empty), n = 3 => 3()^2 * ()
-    result = chain_rule_derivative([], [], 3)
-    assert "3()^2 * ()" == result
 
-def test_chain_rule_derivative_exponent_one():
-    # g(x) = x^3, n = 1 => 1(1x^3)^0 * (3x^2)
-    result = chain_rule_derivative([1], [3], 1)
-    assert "1(" in result
-    assert "1x^3" in result
-    assert ")^0" in result
-    assert "* (3x^2)" in result
-def test_chain_rule_derivative_empty():
-    result = chain_rule_derivative([], [], 2)
-    assert result == "2()^1 * ()"
 
-def test_chain_rule_derivative_float_coeffs():
-    result = chain_rule_derivative([2.5], [2.0], 4)
-    assert "4(" in result
-    assert "2.5x^2" in result
-    assert ")^3" in result
-    assert "* (5.0x^1)" in result
 
-def test_chain_rule_derivative_negative_powers():
-    result = chain_rule_derivative([3], [-2], 2)
-    assert result == "2(3x^-2)^1 * (-6x^-3)"
 
-def test_chain_rule_derivative_exponent_one():
-    result = chain_rule_derivative([4, 1], [3, 0], 1)
-    assert "1(" in result
-    assert "4x^3 + 1x^0" in result
-    assert ")^0" in result
-    assert "* (12x^2)" in result
-def test_chain_rule_derivative_empty():
-    # g(x) = 0, n = 2 => (0)^2 => derivative is 2()^1 * ()
-    result = chain_rule_derivative([], [], 2)
-    assert result == "2()^1 * ()"
 
-def test_chain_rule_derivative_floats():
-    # g(x) = 1.5x^2 + 2.5x, n = 2
-    result = chain_rule_derivative([1.5, 2.5], [2.0, 1.0], 2)
-    assert "2(" in result
-    assert "1.5x^2 + 2.5x^1" in result
-    assert ")^1" in result
-    assert "* (3.0x^1 + 2.5x^0)" in result
 
-def test_chain_rule_derivative_negative_coeffs_powers():
-    # g(x) = -3x^-2, n = 4
-    result = chain_rule_derivative([-3], [-2], 4)
-    assert "4(-3x^-2)^3 * (6x^-3)" in result
 
-def test_chain_rule_derivative_exponent_one():
-    # g(x) = x^2 + 2x, n = 1
-    result = chain_rule_derivative([1, 2], [2, 1], 1)
-    assert "1(1x^2 + 2x^1)^0 * (2x^1 + 2x^0)" in result
-def test_chain_rule_derivative_float_coefficients():
-    # g(x) = 1.5x^2, n = 2 => derivative is 2(1.5x^2)^1 * (3.0x^1)
-    result = chain_rule_derivative([1.5], [2], 2)
-    assert "2(" in result
-    assert "1.5x^2" in result
-    assert ")^1" in result
-    assert "* (3.0x^1)" in result
 
-def test_chain_rule_derivative_negative_powers_and_coeffs():
-    # g(x) = -2x^-3, n = 4 => derivative is 4(-2x^-3)^3 * (6x^-4)
-    result = chain_rule_derivative([-2], [-3], 4)
-    assert "4(" in result
-    assert "-2x^-3" in result
-    assert ")^3" in result
-    assert "* (6x^-4)" in result
 
-def test_chain_rule_derivative_empty_inner():
-    # g(x) = empty, n = 3 => derivative is 3()^2 * ()
-    result = chain_rule_derivative([], [], 3)
-    assert result == "3()^2 * ()"
-def test_chain_rule_derivative_exponent_one():
-    # g(x) = x^2, n = 1 => (x^2)^1 => derivative is 1(x^2)^0 * (2x)
-    result = chain_rule_derivative([1], [2], 1)
-    assert "1(1x^2)^0" in result
-    assert "* (2x^1)" in result
 
-def test_chain_rule_derivative_float_coefficients():
-    # g(x) = 1.5x^2, n = 2 => derivative is 2(1.5x^2)^1 * (3.0x^1)
-    result = chain_rule_derivative([1.5], [2], 2)
-    assert "2(1.5x^2)^1" in result
-    assert "* (3.0x^1)" in result
 
-def test_chain_rule_derivative_negative_powers():
-    # g(x) = -2x^-3, n = 4 => derivative is 4(-2x^-3)^3 * (6x^-4)
-    result = chain_rule_derivative([-2], [-3], 4)
-    assert "4(-2x^-3)^3" in result
-    assert "* (6x^-4)" in result
 
-def test_chain_rule_derivative_empty_inner():
-    # empty inner polynomial
-    result = chain_rule_derivative([], [], 2)
-    assert "2()^1 * ()" in result
+
+
+
+
 
 def test_integrate_cos_basic():
     # Integral of cos(x) from 0 to pi/2 should be sin(pi/2) - sin(0) = 1 - 0 = 1
@@ -612,59 +491,6 @@ class TestDifferentiatePolynomial(unittest.TestCase):
         )
 
 
-from Math.Calculus.Differentiation.chain_rule import format_polynomial as direct_format_polynomial_chain_rule
-class TestChainRuleFormatPolynomial(unittest.TestCase):
-    def test_format_polynomial_single_term(self):
-        self.assertEqual(direct_format_polynomial_chain_rule([4], [2]), "4x^2")
-
-    def test_format_polynomial_zero_coefficient(self):
-        self.assertEqual(direct_format_polynomial_chain_rule([0, 5], [2, 1]), "0x^2 + 5x^1")
-
-    def test_format_polynomial_fractional_power_cast_to_int(self):
-        self.assertEqual(direct_format_polynomial_chain_rule([3], [2.9]), "3x^2")
-
-    def test_format_polynomial_zero_power(self):
-        self.assertEqual(direct_format_polynomial_chain_rule([7], [0]), "7x^0")
-
-    def test_format_polynomial_mismatched_lengths(self):
-        # zip will truncate to the shorter list
-        self.assertEqual(direct_format_polynomial_chain_rule([1, 2], [3]), "1x^3")
-        self.assertEqual(direct_format_polynomial_chain_rule([1], [3, 2]), "1x^3")
-
-    def test_format_polynomial_all_zeros(self):
-        self.assertEqual(direct_format_polynomial_chain_rule([0, 0], [0, 0]), "0x^0 + 0x^0")
-class TestFormatPolynomialChainRule:
-    def test_basic(self):
-        assert format_polynomial_chain_rule([1, 2, 3], [2, 1, 0]) == "1x^2 + 2x^1 + 3x^0"
-
-    def test_floats(self):
-        assert format_polynomial_chain_rule([1.5, 2.5], [2.0, 1.0]) == "1.5x^2 + 2.5x^1"
-
-    def test_empty(self):
-        assert format_polynomial_chain_rule([], []) == ""
-
-    def test_negative_powers(self):
-        assert format_polynomial_chain_rule([5], [-2]) == "5x^-2"
-
-    def test_negative_coeffs(self):
-        assert format_polynomial_chain_rule([-3, -4], [2, 1]) == "-3x^2 + -4x^1"
-
-    def test_zero_coeffs(self):
-        assert format_polynomial_chain_rule([0], [2]) == "0x^2"
-
-    def test_zero_powers(self):
-        assert format_polynomial_chain_rule([5], [0]) == "5x^0"
-
-    def test_float_powers(self):
-        # As per the code, int(power) is used in formatting
-        assert format_polynomial_chain_rule([2], [2.7]) == "2x^2"
-
-    def test_single_term(self):
-        assert format_polynomial_chain_rule([4], [3]) == "4x^3"
-
-    def test_mismatched_lengths(self):
-        # zip will truncate to the shortest list
-        assert format_polynomial_chain_rule([1, 2], [3]) == "1x^3"
 
 if __name__ == '__main__':
     unittest.main()
@@ -746,49 +572,13 @@ def test_quotient_rule_derivative_zero_power_numerator_and_denominator():
     # u' = 0, v' = 0
     result = quotient_rule_derivative([1], [0], [1], [0])
     assert result == "((0) * (1x^0) - (1x^0) * (0)) / (1x^0)^2"
-def test_format_polynomial_chain_rule_basic():
-    coeffs = [3, 5]
-    powers = [2, 1]
-    result = format_polynomial_chain_rule(coeffs, powers)
-    terms = [t.strip() for t in result.split("+")]
-    assert "3x^2" in terms
-    assert "5x^1" in terms
-    assert len(terms) == 2
 
-def test_format_polynomial_chain_rule_empty():
-    result = format_polynomial_chain_rule([], [])
-    assert result == ""
 
-def test_format_polynomial_chain_rule_mixed():
-    result = format_polynomial_chain_rule([-2, 4.5], [3, 0])
-    terms = [t.strip() for t in result.split("+")]
-    assert "-2x^3" in terms
-    assert "4.5x^0" in terms
-    assert len(terms) == 2
 
-def test_format_polynomial_chain_rule_zero_coeffs():
-    assert format_polynomial_chain_rule([0, 0], [2, 1]) == "0x^2 + 0x^1"
 
-def test_format_polynomial_chain_rule_float_powers():
-    assert format_polynomial_chain_rule([2, 3], [2.5, 1.2]) == "2x^2 + 3x^1"
 
-def test_format_polynomial_chain_rule_zero_power():
-    assert format_polynomial_chain_rule([4], [0]) == "4x^0"
 
-def test_format_polynomial_chain_rule_mixed_types():
-    assert format_polynomial_chain_rule([1.5, 2], [3.8, 0]) == "1.5x^3 + 2x^0"
-def test_format_polynomial_chain_rule_single_term():
-    result = format_polynomial_chain_rule([7], [3])
-    terms = [t.strip() for t in result.split("+")]
-    assert "7x^3" in terms
-    assert len(terms) == 1
 
-def test_format_polynomial_chain_rule_zero_coeff():
-    result = format_polynomial_chain_rule([0, 5], [2, 1])
-    terms = [t.strip() for t in result.split("+")]
-    assert "0x^2" in terms
-    assert "5x^1" in terms
-    assert len(terms) == 2
 
 class TestQuotientRule(unittest.TestCase):
     def test_format_polynomial_quotient_rule_basic(self):
@@ -894,236 +684,26 @@ def test_format_polynomial_integration_zero_coefficients():
     assert "0.0x^2" in terms
     assert "0.0x" in terms
     assert "C" in terms
-def test_format_polynomial_chain_rule_zero_coeff():
-    assert format_polynomial_chain_rule([0, 2], [2, 1]) == "0x^2 + 2x^1"
-
-def test_format_polynomial_chain_rule_float_power():
-    assert format_polynomial_chain_rule([3], [2.5]) == "3x^2"
-
-def test_format_polynomial_chain_rule_unequal_lengths():
-    assert format_polynomial_chain_rule([3, 2, 1], [2, 1]) == "3x^2 + 2x^1"
-
-def test_format_polynomial_chain_rule_2_basic():
-    result = format_polynomial_chain_rule([1, 2, 3], [2, 1, 0])
-    terms = [t.strip() for t in result.split("+")]
-    assert "1x^2" in terms
-    assert "2x^1" in terms
-    assert "3x^0" in terms
-    assert len(terms) == 3
-
-def test_format_polynomial_chain_rule_2_floats():
-    result = format_polynomial_chain_rule([1.5, 2.5], [2.0, 1.0])
-    terms = [t.strip() for t in result.split("+")]
-    assert "1.5x^2" in terms
-    assert "2.5x^1" in terms
-    assert len(terms) == 2
-
-class TestFormatPolynomialChainRule:
-    def test_basic_polynomial(self):
-        result = format_polynomial_chain_rule([3, 5], [2, 1])
-        terms = [t.strip() for t in result.split("+")]
-        assert "3x^2" in terms
-        assert "5x^1" in terms
-        assert len(terms) == 2
-
-    def test_empty_polynomial(self):
-        result = format_polynomial_chain_rule([], [])
-        assert result == ""
-
-def test_format_polynomial_chain_rule_2_negative_powers():
-    result = format_polynomial_chain_rule([5], [-2])
-    terms = [t.strip() for t in result.split("+")]
-    assert "5x^-2" in terms
-    assert len(terms) == 1
-
-def test_format_polynomial_chain_rule_2_negative_coeffs():
-    assert format_polynomial_chain_rule([-3, -4], [2, 1]) == "-3x^2 + -4x^1"
-
-def test_format_polynomial_chain_rule_zero_coeff():
-    assert format_polynomial_chain_rule([0, 2], [2, 1]) == "0x^2 + 2x^1"
-
-def test_format_polynomial_chain_rule_zero_power():
-    assert format_polynomial_chain_rule([3], [0]) == "3x^0"
-
-def test_format_polynomial_chain_rule_mismatched_lengths():
-    # zip should truncate to the shortest list
-    assert format_polynomial_chain_rule([1, 2, 3], [2, 1]) == "1x^2 + 2x^1"
-    assert format_polynomial_chain_rule([1, 2], [2, 1, 0]) == "1x^2 + 2x^1"
-
-    result = format_polynomial_chain_rule([-3, -4], [2, 1])
-    terms = [t.strip() for t in result.split("+")]
-    assert "-3x^2" in terms
-    assert "-4x^1" in terms
-    assert len(terms) == 2
-def test_format_polynomial_chain_rule_floating_point_powers():
-    # Floating point powers should be cast to int as per the code: `int(power)`
-    assert format_polynomial_chain_rule([2, 3], [2.5, 1.9]) == "2x^2 + 3x^1"
-
-def test_format_polynomial_chain_rule_zero_coefficients():
-    # Test with 0 as coefficient
-    assert format_polynomial_chain_rule([0, 1], [2, 1]) == "0x^2 + 1x^1"
-
-def test_format_polynomial_chain_rule_zero_power():
-    # Test with 0 as power
-    assert format_polynomial_chain_rule([5, 2], [1, 0]) == "5x^1 + 2x^0"
-
-def test_format_polynomial_chain_rule_unequal_lengths():
-    # zip should truncate to the shorter list
-    assert format_polynomial_chain_rule([1, 2, 3], [1, 0]) == "1x^1 + 2x^0"
-    assert format_polynomial_chain_rule([1, 2], [2, 1, 0]) == "1x^2 + 2x^1"
-    assert format_polynomial_chain_rule([0], [2]) == "0x^2"
-
-def test_format_polynomial_chain_rule_single_term():
-    assert format_polynomial_chain_rule([5], [3]) == "5x^3"
-
-def test_format_polynomial_chain_rule_float_power_cast():
-    assert format_polynomial_chain_rule([2], [3.9]) == "2x^3"
-
-def test_format_polynomial_chain_rule_zero_power():
-    assert format_polynomial_chain_rule([7], [0]) == "7x^0"
-
-    assert format_polynomial_chain_rule([5], [0]) == "5x^0"
-
-def test_format_polynomial_chain_rule_float_power_truncation():
-    assert format_polynomial_chain_rule([1, 2], [2.9, 1.1]) == "1x^2 + 2x^1"
-
-def test_format_polynomial_chain_rule_mismatched_lengths():
-    assert format_polynomial_chain_rule([1, 2, 3], [2, 1]) == "1x^2 + 2x^1"
-    assert format_polynomial_chain_rule([1], [2, 1]) == "1x^2"
-
-def test_format_polynomial_chain_rule_negative_float():
-    assert format_polynomial_chain_rule([-1.5, -2.5], [2, 1]) == "-1.5x^2 + -2.5x^1"
-    def test_mixed_polynomial(self):
-        result = format_polynomial_chain_rule([-2, 4.5], [3, 0])
-        terms = [t.strip() for t in result.split("+")]
-        assert "-2x^3" in terms
-        assert "4.5x^0" in terms
-        assert len(terms) == 2
-
-    def test_floats(self):
-        result = format_polynomial_chain_rule([1.5, 2.5], [2.0, 1.0])
-        terms = [t.strip() for t in result.split("+")]
-        assert "1.5x^2" in terms
-        assert "2.5x^1" in terms
-        assert len(terms) == 2
-
-    def test_negative_powers(self):
-        result = format_polynomial_chain_rule([5], [-2])
-        terms = [t.strip() for t in result.split("+")]
-        assert "5x^-2" in terms
-        assert len(terms) == 1
-
-    def test_negative_coeffs(self):
-        result = format_polynomial_chain_rule([-3, -4], [2, 1])
-        terms = [t.strip() for t in result.split("+")]
-        assert "-3x^2" in terms
-        assert "-4x^1" in terms
-        assert len(terms) == 2
-
-class TestComputePolynomialDerivative:
-    def test_basic_polynomial(self):
-        # f(x) = 3x^2 + 2x^1
-        # f'(x) = 6x^1 + 2x^0
-        coeffs = [3, 2]
-        powers = [2, 1]
-        expected_coeffs = [6, 2]
-        expected_powers = [1, 0]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
-
-    def test_constant_term(self):
-        # f(x) = 5x^0
-        # f'(x) = 0
-        coeffs = [5]
-        powers = [0]
-        expected_coeffs = []
-        expected_powers = []
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
-
-    def test_mixed_terms_with_constant(self):
-        # f(x) = 4x^3 + 2x^0 + x^1
-        # f'(x) = 12x^2 + 1x^0
-        coeffs = [4, 2, 1]
-        powers = [3, 0, 1]
-        expected_coeffs = [12, 1]
-        expected_powers = [2, 0]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
-
-    def test_negative_powers(self):
-        # f(x) = 2x^-2
-        # f'(x) = -4x^-3
-        coeffs = [2]
-        powers = [-2]
-        expected_coeffs = [-4]
-        expected_powers = [-3]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
-
-    def test_fractional_powers(self):
-        # f(x) = 4x^0.5
-        # f'(x) = 2.0x^-0.5
-        coeffs = [4]
-        powers = [0.5]
-        expected_coeffs = [2.0]
-        expected_powers = [-0.5]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
-
-    def test_float_coefficients(self):
-        # f(x) = 1.5x^2
-        # f'(x) = 3.0x^1
-        coeffs = [1.5]
-        powers = [2]
-        expected_coeffs = [3.0]
-        expected_powers = [1]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
 
 
-    def test_zero_coefficients(self):
-        # f(x) = 0x^2
-        # f'(x) = 0x^1
-        coeffs = [0]
-        powers = [2]
-        expected_coeffs = [0]
-        expected_powers = [1]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
 
-    def test_negative_coefficients(self):
-        # f(x) = -3x^2 - 2x^1
-        # f'(x) = -6x^1 - 2x^0
-        coeffs = [-3, -2]
-        powers = [2, 1]
-        expected_coeffs = [-6, -2]
-        expected_powers = [1, 0]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
 
-    def test_multiple_zero_powers(self):
-        # f(x) = 5 + 3
-        # f'(x) = 0
-        coeffs = [5, 3]
-        powers = [0, 0]
-        expected_coeffs = []
-        expected_powers = []
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
 
-    def test_mismatched_lengths(self):
-        # f(x) = 3x^2 + 2x^1 (powers list has an extra element)
-        coeffs = [3, 2]
-        powers = [2, 1, 0]
-        expected_coeffs = [6, 2]
-        expected_powers = [1, 0]
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
 
-        # coeffs list has an extra element
-        coeffs2 = [3, 2, 1]
-        powers2 = [2, 1]
-        assert compute_polynomial_derivative(coeffs2, powers2) == (expected_coeffs, expected_powers)
 
-    def test_empty_polynomial(self):
-        # f(x) = 0
-        # f'(x) = 0
-        coeffs = []
-        powers = []
-        expected_coeffs = []
-        expected_powers = []
-        assert compute_polynomial_derivative(coeffs, powers) == (expected_coeffs, expected_powers)
+
+
+
+
+
+
+
+
+
+
+
+
+
 class TestCalculusDifferentiation(unittest.TestCase):
     def test_format_polynomial_product_rule_basic(self):
         result = format_polynomial_product_rule([2, 3], [1, 2])
@@ -1259,6 +839,28 @@ class TestTrigIntegration:
         """Test integral of cos(x) from a to a = 0"""
         result = integrate_cos(math.pi, math.pi)
         assert math.isclose(result, 0.0, abs_tol=1e-5)
+
+    def test_integrate_cos_reversed_bounds(self):
+        result = integrate_cos(math.pi / 2, 0)
+        assert math.isclose(result, -1.0, rel_tol=1e-5)
+
+    def test_integrate_cos_fractional_bounds(self):
+        result = integrate_cos(math.pi / 6, math.pi / 3)
+        expected = math.sin(math.pi / 3) - math.sin(math.pi / 6)
+        assert math.isclose(result, expected, rel_tol=1e-5)
+
+    def test_integrate_cos_same_bounds(self):
+        result = integrate_cos(math.pi, math.pi)
+        assert math.isclose(result, 0.0, abs_tol=1e-9)
+
+    def test_integrate_cos_half_period(self):
+        result = integrate_cos(0, math.pi)
+        assert math.isclose(result, 0.0, abs_tol=1e-9)
+
+    def test_integrate_cos_multiple_periods(self):
+        result = integrate_cos(-2 * math.pi, 4 * math.pi)
+        assert math.isclose(result, 0.0, abs_tol=1e-9)
+
 def test_quotient_rule_derivative_happy_path():
     # Normal polynomial differentiation: u(x) = 2x^2, v(x) = 3x^1
     # u' = 4x^1, v' = 3x^0
@@ -1289,64 +891,6 @@ def test_quotient_rule_derivative_zero_polynomials_edge():
     assert result == "((0) * (1x^1) - (0x^0) * (1x^0)) / (1x^1)^2"
 
 
-class TestFormatPolynomialChainRule(unittest.TestCase):
-    def test_single_term(self):
-        self.assertEqual(format_polynomial_chain_rule([5], [2]), "5x^2")
-
-    def test_zero_coefficient(self):
-        self.assertEqual(format_polynomial_chain_rule([0], [3]), "0x^3")
-
-    def test_zero_power(self):
-        self.assertEqual(format_polynomial_chain_rule([4], [0]), "4x^0")
-
-    def test_float_power_truncation(self):
-        # int() truncates floats towards zero
-        self.assertEqual(format_polynomial_chain_rule([2], [3.9]), "2x^3")
-
-    def test_negative_power(self):
-        self.assertEqual(format_polynomial_chain_rule([7], [-2]), "7x^-2")
-
-    def test_multiple_terms(self):
-        self.assertEqual(format_polynomial_chain_rule([3, -2, 4.5], [2, 1, 0]), "3x^2 + -2x^1 + 4.5x^0")
-
-    def test_empty_lists(self):
-        self.assertEqual(format_polynomial_chain_rule([], []), "")
-
-    def test_mismatched_list_lengths(self):
-        # zip() truncates to the shortest list
-        self.assertEqual(format_polynomial_chain_rule([1, 2], [3]), "1x^3")
-        self.assertEqual(format_polynomial_chain_rule([1], [3, 2]), "1x^3")
-        """Test integral of cos(x) from -pi to 0 = 0"""
-        result = integrate_cos(-math.pi, 0)
-        assert math.isclose(result, 0.0, abs_tol=1e-9)
-
-    def test_integrate_cos_reversed_bounds(self):
-        """Test integral of cos(x) from pi/2 to 0 = -1"""
-        result = integrate_cos(math.pi / 2, 0)
-        assert math.isclose(result, -1.0, rel_tol=1e-5)
-
-    def test_integrate_cos_fractional_bounds(self):
-        """Test integral of cos(x) from pi/6 to pi/3 = (sqrt(3)/2 - 1/2)"""
-        result = integrate_cos(math.pi / 6, math.pi / 3)
-        expected = math.sin(math.pi / 3) - math.sin(math.pi / 6)
-        assert math.isclose(result, expected, rel_tol=1e-5)
-    def test_integrate_cos_same_bounds(self):
-        """Test integral of cos(x) from pi to pi = 0"""
-        result = integrate_cos(math.pi, math.pi)
-        assert math.isclose(result, 0.0, abs_tol=1e-9)
-
-    def test_integrate_cos_half_period(self):
-        """Test integral of cos(x) from 0 to pi = 0"""
-        result = integrate_cos(0, math.pi)
-        assert math.isclose(result, 0.0, abs_tol=1e-9)
-
-    def test_integrate_cos_multiple_periods(self):
-        """Test integral of cos(x) over multiple periods"""
-        result = integrate_cos(-2 * math.pi, 4 * math.pi)
-        assert math.isclose(result, 0.0, abs_tol=1e-9)
-        """Test integral of cos(x) from -pi/2 to pi/2 = 2"""
-        result = integrate_cos(-math.pi / 2, math.pi / 2)
-        assert math.isclose(result, 2.0, rel_tol=1e-5)
 
 
 class TestProductRuleComputePolynomialDerivativeStr(unittest.TestCase):
@@ -1378,35 +922,3 @@ if __name__ == '__main__':
     unittest.main()
 
 
-class TestFormatPolynomialChainRule(unittest.TestCase):
-    def test_basic_positive_powers(self):
-        result = format_polynomial_chain_rule([1, 2, 3], [2, 1, 0])
-        self.assertEqual(result, "1x^2 + 2x^1 + 3x^0")
-
-    def test_zero_coefficients(self):
-        result = format_polynomial_chain_rule([0, 0], [2, 1])
-        self.assertEqual(result, "0x^2 + 0x^1")
-
-    def test_zero_powers(self):
-        result = format_polynomial_chain_rule([5, 4], [0, 0])
-        self.assertEqual(result, "5x^0 + 4x^0")
-
-    def test_negative_powers(self):
-        result = format_polynomial_chain_rule([5, -2], [-2, -3])
-        self.assertEqual(result, "5x^-2 + -2x^-3")
-
-    def test_negative_coefficients(self):
-        result = format_polynomial_chain_rule([-3, -4], [2, 1])
-        self.assertEqual(result, "-3x^2 + -4x^1")
-
-    def test_floating_point_coefficients(self):
-        result = format_polynomial_chain_rule([1.5, 2.5], [2.0, 1.0])
-        self.assertEqual(result, "1.5x^2 + 2.5x^1")
-
-    def test_empty_lists(self):
-        result = format_polynomial_chain_rule([], [])
-        self.assertEqual(result, "")
-
-    def test_single_element(self):
-        result = format_polynomial_chain_rule([7], [3])
-        self.assertEqual(result, "7x^3")
